pyproject.toml
[project]
name = "margadarsaka"
version = "0.2.0"
description = "AI-powered psychological career advisor specializing in mental, analytical, and deduction skills with Indian cultural awareness"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    # AI and ML libraries
    "google-generativeai>=0.3.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "datasets>=2.14.0",
    # Data handling
    "polars>=0.20.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    # Translation and language
    "deep-translator>=1.11.0",
    "googletrans>=4.0.0",
    # PDF and image generation
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    # Resume analysis
    "pypdf2>=3.0.0",
    "python-docx>=0.8.11",
    "spacy>=3.7.0",
    # Psychological testing
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    # UI enhancements
    "streamlit-option-menu>=0.3.6",
    "streamlit-lottie>=0.0.5",
    "streamlit-authenticator>=0.2.3",
    "seaborn>=0.13.2",
    "doppler-sdk==1.3.0",
    "python-dotenv>=1.1.1",
    "appwrite>=8.0.0",
]
authors = [
    { name = "Dakshx-Gupta", email = "dakshx.gupta@example.com" }
]
license = { file = "LICENSE" }
keywords = ["career", "advisor", "ai", "psychology", "mental-skills", "india", "riasec"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Dakshx-Gupta/Margadarsaka"
Repository = "https://github.com/Dakshx-Gupta/Margadarsaka"
Issues = "https://github.com/Dakshx-Gupta/Margadarsaka/issues"

[project.scripts]
margadarsaka = "margadarsaka.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
